Almost fully working userSearch and artistPage
On every search of a valid Display_name for a user that is in the database, the artistpage album display will change correctly. However, if you first search a valid user, then go back and search an invalid user, the first searched user  albums will continue to display until you search fora different valid user.
For our demonstration we could sidestep this cleverly by searching first for an invalid user, which will correctly display an empty listing,, and THEN search for valid users. but it's still a bug.